module1/prac9/lecture/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
psycopg2-binary==2.9.10
loguru==0.7.3
requests==2.32.5